+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/FMC/A32NX_FMCMainDisplay.js
@@ -61,6 +61,7 @@ class FMCMainDisplay extends BaseAirliners {
         this._vor2Frequency = undefined;
         this._vor2Course = undefined;
         this._ilsFrequency = undefined;
+        this._ilsIcao = undefined;
         this._ilsIdent = undefined;
         this._ilsFrequencyPilotEntered = undefined;
         this._ilsIdentPilotEntered = undefined;
@@ -157,9 +158,12 @@ class FMCMainDisplay extends BaseAirliners {
         this.cruiseFlightLevelTimeOut = undefined;
         this.flaps = undefined;
         this.ths = undefined;
+        this.ilsAutoFrequency = undefined;
+        this.ilsAutoIcao = undefined;
         this.ilsAutoIdent = undefined;
         this.ilsAutoCourse = undefined;
         this.ilsAutoTuned = undefined;
+        this.tempFpPendingAutoTune = undefined;
         this.ilsTakeoffAutoTuned = undefined;
         this.ilsApproachAutoTuned = undefined;
         this.climbTransitionGroundAltitude = undefined;
@@ -286,6 +290,8 @@ class FMCMainDisplay extends BaseAirliners {
                 }
             }
         }, 15000);
+
+        SimVar.SetSimVarValue('L:A32NX_FM_LS_COURSE', 'number', -1);
     }
 
     initVariables() {
@@ -332,10 +338,11 @@ class FMCMainDisplay extends BaseAirliners {
         this._vor2Frequency = 0;
         this._vor2Course = 0;
         this._ilsFrequency = 0;
-        this._ilsIdent = '';
+        this._ilsIcao = undefined;
+        this._ilsIdent = undefined;
         this._ilsFrequencyPilotEntered = false;
         this._ilsIdentPilotEntered = false;
-        this._ilsCourse = 0;
+        this._ilsCourse = undefined;
         this._adf1Frequency = 0;
         this._adf2Frequency = 0;
         this._debug = 0;
@@ -477,11 +484,14 @@ class FMCMainDisplay extends BaseAirliners {
         this.cruiseFlightLevelTimeOut = undefined;
         this.flaps = NaN;
         this.ths = NaN;
-        this.ilsAutoIdent = '';
-        this.ilsAutoCourse = 0;
+        this.ilsAutoFrequency = undefined;
+        this.ilsAutoIcao = undefined;
+        this.ilsAutoIdent = undefined;
+        this.ilsAutoCourse = undefined;
         this.ilsAutoTuned = false;
         this.ilsTakeoffAutoTuned = false;
         this.ilsApproachAutoTuned = false;
+        this.tempFpPendingAutoTune = false;
         this.climbTransitionGroundAltitude = null;
         this.altDestination = undefined;
         this.flightNumber = undefined;
@@ -1651,7 +1661,7 @@ class FMCMainDisplay extends BaseAirliners {
 
     setOriginRunwayIndex(runwayIndex, callback = EmptyCallback.Boolean) {
         this.ensureCurrentFlightPlanIsTemporary(() => {
-            this.clearAutotunedIls();
+            this.tempFpPendingAutoTune = true;
             this.flightPlanManager.setDepartureProcIndex(-1, () => {
                 this.flightPlanManager.setOriginRunwayIndex(runwayIndex, () => {
                     return callback(true);
@@ -1663,7 +1673,7 @@ class FMCMainDisplay extends BaseAirliners {
     setRunwayIndex(runwayIndex, callback = EmptyCallback.Boolean) {
         this.ensureCurrentFlightPlanIsTemporary(() => {
             const routeOriginInfo = this.flightPlanManager.getOrigin().infos;
-            this.clearAutotunedIls();
+            this.tempFpPendingAutoTune = true;
             if (!this.flightPlanManager.getOrigin()) {
                 this.addNewMessage(NXFictionalMessages.noOriginSet);
                 return callback(false);
@@ -1750,7 +1760,7 @@ class FMCMainDisplay extends BaseAirliners {
                         SimVar.SetSimVarValue("L:A32NX_PRESS_AUTO_LANDING_ELEVATION", "feet", A32NX_Util.meterToFeet(runway.elevation));
                     }
                 }
-                this.clearAutotunedIls();
+                this.tempFpPendingAutoTune = true;
                 callback(true);
             }).catch(console.error);
         });
@@ -1763,20 +1773,29 @@ class FMCMainDisplay extends BaseAirliners {
             try {
                 const ils = await this.facilityLoader.getFacility(ilsIcao).catch(console.error);
                 if (ils.infos.frequencyMHz > 1) {
-                    console.log('Auto-tuning ILS', ils);
-                    this.connectIlsFrequency(ils.infos.frequencyMHz);
+                    this.ilsAutoFrequency = ils.infos.frequencyMHz;
+                    this.ilsAutoIcao = ils.infos.icao;
                     this.ilsAutoIdent = ils.infos.ident;
-                    this.ilsAutoCourse = finalLeg.course;
+                    this.ilsAutoCourse = Math.round(finalLeg.course) % 360;
                     this.ilsAutoTuned = true;
+                    if (!this._ilsFrequencyPilotEntered && !this._ilsIdentPilotEntered) {
+                        this.connectIlsFrequency(this.ilsAutoFrequency);
+                    }
+                    if (this.ilsCourse) {
+                        this.checkRunwayLsCourseMismatch();
+                    } else {
+                        await this.updateIlsCourse();
+                    }
                     if (this.currentFlightPhase > FmgcFlightPhases.TAKEOFF) {
                         this.ilsApproachAutoTuned = true;
                     } else {
                         this.ilsTakeoffAutoTuned = true;
                     }
+                    this.checkRunwayLsMismatch();
                     return true;
                 }
             } catch (error) {
-                console.log('tuneIlsFromApproach', error);
+                console.error('tuneIlsFromApproach', error);
                 return false;
             }
         }
@@ -1787,20 +1806,20 @@ class FMCMainDisplay extends BaseAirliners {
         this.ilsAutoTuned = false;
         this.ilsApproachAutoTuned = false;
         this.ilsTakeoffAutoTuned = false;
-        this.ilsAutoIdent = "";
-        this.ilsAutoCourse = 0;
+        this.ilsAutoFrequency = undefined;
+        this.ilsAutoIdent = undefined;
+        this.ilsAutoCourse = undefined;
+        this.updateIlsCourse();
     }
 
     async updateIls() {
-        if (this._ilsFrequencyPilotEntered || this._ilsIdentPilotEntered) {
-            return;
-        }
+        await this.updateIlsCourse();
 
         let airport;
         let runway;
 
         if (this.currentFlightPhase > FmgcFlightPhases.TAKEOFF) {
-            if (this.ilsApproachAutoTuned) {
+            if (this.ilsApproachAutoTuned || this.flightPlanManager.getCurrentFlightPlanIndex() !== 0) {
                 return;
             }
             this.ilsAutoTuned = false;
@@ -1825,7 +1844,7 @@ class FMCMainDisplay extends BaseAirliners {
                 }
             }
         } else {
-            if (this.ilsTakeoffAutoTuned) {
+            if (this.ilsTakeoffAutoTuned || this.flightPlanManager.getCurrentFlightPlanIndex() !== 0) {
                 return;
             }
             this.ilsAutoTuned = false;
@@ -1835,7 +1854,7 @@ class FMCMainDisplay extends BaseAirliners {
 
         // If the airport has correct navdata, the ILS will be listed as the reference navaid (originIcao in MSFS land) on at least the last leg of the
         // ILS approach procedure(s). Tuning this way gives us the ident, and the course
-        if (airport && airport.infos && runway) {
+        if (airport && airport.infos && airport.infos.icao.charAt(0) === 'A' && runway) {
             for (let i = 0; i < airport.infos.approaches.length && !this.ilsAutoTuned; i++) {
                 const appr = airport.infos.approaches[i];
                 // L(eft), C(entre), R(ight), T(true North) are the possible runway designators (ARINC424)
@@ -1852,6 +1871,36 @@ class FMCMainDisplay extends BaseAirliners {
                     }
                 }
             }
+        }
+    }
+
+    async updateIlsCourse() {
+        let course = -1;
+        if (this.ilsCourse !== undefined) {
+            course = this.ilsCourse;
+        } else if (this.ilsAutoTuned && (!this._ilsIdentPilotEntered || this._ilsIcao === this.ilsAutoIcao) && !this._ilsFrequencyPilotEntered) {
+            course = this.ilsAutoCourse;
+        } else if (this.ilsFrequency > 0 && SimVar.GetSimVarValue('NAV HAS LOCALIZER:3', 'boolean') === 1) {
+            course = SimVar.GetSimVarValue('NAV LOCALIZER:3', 'degrees');
+        }
+        return SimVar.SetSimVarValue('L:A32NX_FM_LS_COURSE', 'number', course);
+    }
+
+    checkRunwayLsMismatch() {
+        if (!this.ilsAutoTuned || this.currentFlightPhase === FmgcFlightPhases.DONE) {
+            return false;
+        }
+
+        if ((this._ilsFrequencyPilotEntered && Math.abs(this.ilsFrequency - this.ilsAutoFrequency) >= 0.05) || (this._ilsIdentPilotEntered && this._ilsIcao !== this.ilsAutoIcao)) {
+            this.addNewMessage(NXSystemMessages.rwyLsMismatch);
+        }
+
+        // manually entered course mismatch is handled separately to avoid unwanted messages
+    }
+
+    checkRunwayLsCourseMismatch() {
+        if (this.ilsAutoTuned && !this._ilsFrequencyPilotEntered && (!this._ilsIdentPilotEntered || this._ilsIcao === this.ilsAutoIcao) && this.ilsCourse !== undefined && Math.abs(Avionics.Utils.diffAngle(this.ilsCourse, this.ilsAutoCourse)) > 3) {
+            this.addNewMessage(NXSystemMessages.rwyLsMismatch);
         }
     }
 
@@ -2147,6 +2196,7 @@ class FMCMainDisplay extends BaseAirliners {
         this.flightPlanManager.setCurrentFlightPlanIndex(0, () => {
             SimVar.SetSimVarValue("L:FMC_FLIGHT_PLAN_IS_TEMPORARY", "number", 0);
             SimVar.SetSimVarValue("L:MAP_SHOW_TEMPORARY_FLIGHT_PLAN", "number", 0);
+            this.tempFpPendingAutoTune = false;
             callback();
         });
     }
@@ -2157,6 +2207,10 @@ class FMCMainDisplay extends BaseAirliners {
                 this.flightPlanManager.setCurrentFlightPlanIndex(0, () => {
                     SimVar.SetSimVarValue("L:FMC_FLIGHT_PLAN_IS_TEMPORARY", "number", 0);
                     SimVar.SetSimVarValue("L:MAP_SHOW_TEMPORARY_FLIGHT_PLAN", "number", 0);
+                    if (this.tempFpPendingAutoTune) {
+                        this.clearAutotunedIls();
+                        this.tempFpPendingAutoTune = false;
+                    }
                     callback();
                 });
             }).catch(console.error);
@@ -3201,22 +3255,42 @@ class FMCMainDisplay extends BaseAirliners {
 
     setIlsFrequency(s, callback) {
         if (s === FMCMainDisplay.clrValue) {
-            this.ilsFrequency = 0;
-            this.radioNav.setILSActiveFrequency(1, 0);
+            if (!this._ilsIdentPilotEntered && !this._ilsFrequencyPilotEntered) {
+                this.addNewMessage(NXSystemMessages.notAllowed);
+                return callback(false);
+            }
+
             this._ilsFrequencyPilotEntered = false;
             this._ilsIdentPilotEntered = false;
-            return callback(true);
+            this._ilsIcao = undefined;
+            if (this.ilsAutoTuned) {
+                this.connectIlsFrequency(this.ilsAutoFrequency);
+                this.checkRunwayLsCourseMismatch();
+            } else {
+                this.ilsCourse = undefined;
+                this.ilsFrequency = 0;
+                this.radioNav.setILSActiveFrequency(1, 0);
+            }
+            this.updateIlsCourse().then(() => {
+                callback(true);
+            });
+            return;
         }
         if (s.match(/^[0-9]{3}(\.[0-9]{1,2})$/) !== null) {
             const v = parseFloat(s);
             const freq = Math.round(v * 100) / 100;
 
             if (this.connectIlsFrequency(freq)) {
+                this._ilsIcao = undefined;
                 this._ilsIdent = undefined;
                 this._ilsFrequencyPilotEntered = true;
                 this._ilsIdentPilotEntered = false;
-                this.clearAutotunedIls();
-                return callback(true);
+                this.ilsCourse = undefined;
+                this.checkRunwayLsMismatch();
+                this.updateIlsCourse().then(() => {
+                    callback(true);
+                });
+                return;
             } else {
                 this.addNewMessage(NXSystemMessages.entryOutOfRange);
                 return callback(false);
@@ -3225,11 +3299,18 @@ class FMCMainDisplay extends BaseAirliners {
             this.getOrSelectILSsByIdent(s, (navaid) => {
                 if (navaid) {
                     if (this.connectIlsFrequency(Math.round(navaid.infos.frequencyMHz * 100) / 100)) {
+                        this._ilsIcao = navaid.infos.icao;
                         this._ilsIdent = s;
                         this._ilsFrequencyPilotEntered = false;
                         this._ilsIdentPilotEntered = true;
-                        this.clearAutotunedIls();
-                        return callback(true);
+                        if (!this.ilsAutoTuned || this.ilsAutoIcao !== this._ilsIcao) {
+                            this.ilsCourse = undefined;
+                        }
+                        this.checkRunwayLsMismatch();
+                        this.updateIlsCourse().then(() => {
+                            callback(true);
+                        });
+                        return;
                     } else {
                         this.addNewMessage(NXSystemMessages.databaseCodingError);
                         return callback(false);
@@ -3244,6 +3325,52 @@ class FMCMainDisplay extends BaseAirliners {
             this.addNewMessage(NXSystemMessages.formatError);
             return callback(false);
         }
+    }
+
+    setLsCourse(s, callback) {
+        if (!this.ilsAutoTuned && !this._ilsFrequencyPilotEntered && !this._ilsIdentPilotEntered) {
+            this.addNewMessage(NXSystemMessages.notAllowed);
+            return callback(false);
+        }
+
+        if (s === FMCMainDisplay.clrValue) {
+            if (this.ilsCourse !== undefined) {
+                this.ilsCourse = undefined;
+                this.updateIlsCourse().then(() => {
+                    callback(true);
+                });
+                return;
+            } else {
+                this.addNewMessage(NXSystemMessages.notAllowed);
+                return callback(false);
+            }
+        }
+
+        const m = s.match(/^(F|B)?([0-9]{1,3})(F|B)?$/);
+        const direction = m[1] || m[3];
+        const course = parseInt(m[2]);
+        if (m !== null && direction !== undefined && (m[1] === undefined || m[3] === undefined)) {
+            if (course > 360) {
+                this.addNewMessage(NXSystemMessages.entryOutOfRange);
+                return callback(false);
+            }
+            if (direction === 'B') {
+                this.addNewMessage(NXFictionalMessages.notYetImplemented);
+                return callback(false);
+            }
+
+            this.ilsCourse = course % 360;
+
+            this.checkRunwayLsCourseMismatch();
+
+            this.updateIlsCourse().then(() => {
+                callback(true);
+            });
+            return;
+        }
+
+        this.addNewMessage(NXSystemMessages.formatError);
+        return callback(false);
     }
 
     initRadioNav(_boot) {
@@ -3403,8 +3530,8 @@ class FMCMainDisplay extends BaseAirliners {
         this._ilsFrequency = _frq;
     }
 
-    set ilsCourse(_crs) {
-        this._ilsCourse = _crs;
+    set ilsCourse(crs) {
+        this._ilsCourse = crs;
     }
 
     set adf1Frequency(_frq) {