+++ .github/CHANGELOG.md
@@ -94,6 +94,7 @@
 1. [FM] Stored waypoints - @tracernz (Mike)
 1. [ATSU] Import lat/lon waypoints correctly from SimBrief - @tracernz (Mike)
 1. [MCDU] Added manual ILS tuning by ident on RAD NAV page - @tracernz (Mike)
+1. [MCDU] Support manual entry of forward LS course - @tracernz (Mike)
 
 ## 0.7.0